Authorization added

+++ backend/controllers/goalController.js
@@ -1,10 +1,11 @@
 const asyncHanlder = require('express-async-handler')
-const Goal = require('../models/goalModel')
+const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 // @desc Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHanlder( async (req,res) =>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals);
 });
 
@@ -19,7 +20,8 @@ const setGoals = asyncHanlder( async (req,res) =>{
     }
 
     const goal = await Goal.create({
-        text:req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
@@ -35,6 +37,20 @@ const updateGoal = asyncHanlder( async (req,res) =>{
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the logged in user matches the goal user
+    if( goal.user.toString() !== user.id ){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{ new:true })
@@ -54,6 +70,21 @@ const deleteGoal = asyncHanlder( async (req,res) =>{
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the logged in user matches the goal user
+    if( goal.user.toString() !== user.id ){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
+    
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id,)
 
     res.status(200).json(deletedGoal)
